+++ changelog.md
@@ -5,7 +5,7 @@ failed cypress tests
 - will sort invoices by datefailed
 - will multi-sort clientsfailed
 # pending requested features
-- last invoices in client profile. currently is showing unsorted invoices grid.
+- test last invoices in client profile.
 # pending features
 2. some unit testing 
 4. cypress extra